+++ .air.toml
@@ -1,0 +1,8 @@
+root = "."
+tmp_dir = "tmp"
+
+[build]
+cmd = "go build -o ./tmp/main.exe ./src"
+bin = "./tmp/main.exe"
+include_ext = ["go"]
+exclude_dir = ["tmp", "vendor"]
