+++ .env.example
@@ -1,4 +1,5 @@
 ADMIN_USERNAME=youradmin
 ADMIN_PASSWORD=sha256_hash_of_password
+PRIVILEGE_PASSWORD=your_password_here
 TURSO_DATABASE_URL=turso_url
 TURSO_AUTH_TOKEN=auth_token
