+++ .gitignore
@@ -20,3 +20,7 @@ Thumbs.db
 .vscode/
 .idea/
 *.bak
+
+todo.md
+
+tmp
