+++ templates/maintainer_login.tmpl
@@ -1,0 +1,99 @@
+<!DOCTYPE html>
+<html lang="en">
+<head>
+    <meta charset="UTF-8">
+    <meta name="viewport" content="width=device-width, initial-scale=1.0">
+    <title>Maintainer Login</title>
+    <style>
+        body {
+            background: #181a20;
+            color: #eaeaea;
+            font-family: 'Segoe UI', 'Roboto', 'Arial', sans-serif;
+            margin: 0;
+            padding: 0;
+            display: flex;
+            justify-content: center;
+            align-items: center;
+            min-height: 100vh;
+        }
+        .login-container {
+            background: #23272f;
+            border-radius: 18px;
+            box-shadow: 0 4px 32px #000a;
+            padding: 32px;
+            width: 360px;
+        }
+        h2 {
+            text-align: center;
+            margin-top: 0;
+            margin-bottom: 24px;
+            font-size: 2rem;
+            color: #aa3fff;
+            letter-spacing: 1px;
+            text-shadow: 0 2px 8px rgba(170, 63, 255, 0.3);
+        }
+        .form-group {
+            margin-bottom: 20px;
+        }
+        label {
+            display: block;
+            margin-bottom: 8px;
+            color: #17b897;
+            font-weight: bold;
+        }
+        input[type="text"], input[type="password"] {
+            width: 100%;
+            padding: 12px;
+            background: #32363f;
+            border: 1px solid #444;
+            border-radius: 6px;
+            color: white;
+            font-size: 1rem;
+            box-sizing: border-box;
+        }
+        input[type="text"]:focus, input[type="password"]:focus {
+            border-color: #aa3fff;
+            outline: none;
+        }
+        input[type="submit"] {
+            width: 100%;
+            padding: 12px;
+            background: #aa3fff;
+            color: white;
+            border: none;
+            border-radius: 6px;
+            font-size: 1rem;
+            font-weight: bold;
+            cursor: pointer;
+            transition: background 0.2s;
+            margin-top: 12px;
+        }
+        input[type="submit"]:hover {
+            background: #be6fff;
+        }
+        .error {
+            background: rgba(255, 59, 59, 0.1);
+            color: #ff5b5b;
+            padding: 12px;
+            border-radius: 6px;
+            margin-bottom: 20px;
+            text-align: center;
+        }
+    </style>
+</head>
+<body>
+    <div class="login-container">
+        <h2>Maintainer Login</h2>
+        {{if .error}}
+            <div class="error">{{.error}}</div>
+        {{end}}
+        <form action="/maintainer/login" method="post">
+            <div class="form-group">
+                <label for="password">Password</label>
+                <input type="password" name="password" id="password" required autofocus>
+            </div>
+            <input type="submit" value="Login">
+        </form>
+    </div>
+</body>
+</html>
