+++ templates/maintainer_users.tmpl
@@ -1,0 +1,158 @@
+{{ define "maintainer_users.tmpl" }}
+<!DOCTYPE html>
+<html>
+<head>
+    <meta charset="UTF-8">
+    <meta name="viewport" content="width=device-width, initial-scale=1.0">
+    <title>Maintainer - Manage Past Users</title>
+    <style>
+        body {
+            background: #181a20;
+            color: #eaeaea;
+            font-family: 'Segoe UI', 'Roboto', 'Arial', sans-serif;
+            margin: 0;
+            padding: 0;
+        }
+        .container {
+            max-width: 950px;
+            margin: 40px auto;
+            background: #23272f;
+            border-radius: 18px;
+            box-shadow: 0 4px 32px #000a;
+            padding: 32px;
+        }
+        h1, h2 {
+            text-align: center;
+            letter-spacing: 1px;
+        }
+        h1 {
+            margin-top: 16px;
+            font-size: 2.2rem;
+            color: #aa3fff;
+            text-shadow: 0 2px 8px rgba(170, 63, 255, 0.3);
+        }
+        h2 {
+            margin: 32px 0 16px;
+            font-size: 1.5rem;
+            color: #17b897;
+        }
+        .form-container {
+            background: #2a2e3a;
+            border-radius: 12px;
+            padding: 24px;
+            margin: 24px 0;
+        }
+        input[type="text"] {
+            background: #32363f;
+            border: none;
+            padding: 12px 16px;
+            margin: 0 8px 12px 0;
+            border-radius: 6px;
+            color: white;
+            font-size: 1rem;
+        }
+            input[type="number"] {
+            background: #32363f;
+            border: none;
+            padding: 12px 16px;
+            margin: 0 8px 12px 0;
+            border-radius: 6px;
+            color: white;
+            font-size: 1rem;
+        }
+        button {
+            background: #aa3fff;
+            color: white;
+            border: none;
+            padding: 12px 24px;
+            border-radius: 6px;
+            cursor: pointer;
+            font-weight: bold;
+            transition: all 0.2s;
+        }
+        button:hover {
+            background: #be6fff;
+        }
+        table {
+            width: 100%;
+            border-collapse: collapse;
+            margin: 24px 0;
+        }
+        th, td {
+            padding: 12px 16px;
+            text-align: left;
+            border-bottom: 1px solid #32363f;
+        }
+        th {
+            background: #32363f;
+            color: #17b897;
+            font-weight: bold;
+        }
+        tr:nth-child(even) {
+            background: #20232a;
+        }
+        tr:hover {
+            background: #282c34;
+        }
+        .delete-btn {
+            background:rgb(255, 0, 0);
+            padding: 8px 16px;
+        }
+        .delete-btn:hover {
+            background:rgb(190, 0, 0);
+        }
+        .back-link {
+            display: inline-block;
+            margin-top: 24px;
+            color: #aa3fff;
+            text-decoration: none;
+            font-weight: bold;
+            transition: color 0.2s;
+        }
+        .back-link:hover {
+            color: #be6fff;
+        }
+    </style>
+</head>
+<body>
+    <div class="container">
+        <h1>Manage Codeforces Past Users</h1>
+        <div class="form-container">
+            <form method="POST" action="/maintainer/users/add">
+                <input type="text" name="handle" placeholder="Codeforces Handle" required>
+                <input type="text" name="display_name" placeholder="Display Name" required>
+                <input type="number" name="batch" placeholder="Batch Year (e.g. 2023)" required>
+                <button type="submit">Add User</button>
+            </form>
+        </div>
+        <h2>Tracked Users</h2>
+    <table>
+    <tr>
+        <th>ID</th>
+        <th>Handle</th>
+        <th>Display Name</th>
+        <th>Batch</th>
+        <th>Action</th>
+    </tr>
+
+    {{ range .past_users }}
+    <tr>
+        <td>{{ .id }}</td>
+        <td>{{ .handle }}</td>
+        <td>{{ .display_name }}</td>
+        <td>{{ .batch }}</td>
+
+        <td>
+            <form method="POST" action="/maintainer/users/delete" style="display:inline;">
+                <input type="hidden" name="id" value="{{ .id }}">
+                <button type="submit" class="delete-btn">Delete</button>
+            </form>
+        </td>
+    </tr>
+    {{ end }}
+</table>
+        <a href="/maintainer/dashboard" class="back-link">← Back to Dashboard</a>
+    </div>
+</body>
+</html>
+{{ end }}
